refactor: one carousel container with two carousel component
